--- test/unit/CPAC/pipeline/subject_list_4_unittest.yml
@@ -1,37 +1,0 @@
--
-    subject_id: '0010128'
-    unique_id: 'session_1'
-    anat: '/home/rtungaraza/CPAC_testData/data/unsorted/0010128/session_1/anat_1/mprage.nii.gz'
-    rest: 
-      rest_1_rest: '/home/rtungaraza/CPAC_testData/data/unsorted/0010128/session_1/rest_1/rest.nii.gz'
-    scan_parameters:
-        tr: '2.5'
-        acquisition: 'seq+z'
-        reference: '24'
-        first_tr: ''
-        last_tr: ''
--
-    subject_id: '0021019'
-    unique_id: 'session_1'
-    anat: '/home/rtungaraza/CPAC_testData/data/unsorted/0021019/session_1/anat_1/mprage.nii.gz'
-    rest: 
-      rest_1_rest: '/home/rtungaraza/CPAC_testData/data/unsorted/0021019/session_1/rest_1/rest.nii.gz'
-    scan_parameters:
-        tr: '2.5'
-        acquisition: 'seq+z'
-        reference: '24'
-        first_tr: ''
-        last_tr: ''
--
-    subject_id: '3154996'
-    unique_id: 'session_1'
-    anat: '/home/rtungaraza/CPAC_testData/data/unsorted/3154996/session_1/anat_1/mprage.nii.gz'
-    rest: 
-      rest_1_rest: '/home/rtungaraza/CPAC_testData/data/unsorted/3154996/session_1/rest_1/rest.nii.gz'
-    scan_parameters:
-        tr: '2.5'
-        acquisition: 'seq+z'
-        reference: '24'
-        first_tr: ''
-        last_tr: ''
-
